LocalProfile moved to 'local.profile' package
