+++ src/main/java/egl/client/repository/profile/LocalProfileRepository.java
@@ -1,6 +1,6 @@
 package egl.client.repository.profile;
 
-import egl.client.model.profile.LocalProfile;
+import egl.client.model.local.profile.LocalProfile;
 import egl.client.repository.EntityRepository;
 import org.springframework.stereotype.Repository;
