+++ src/main/java/egl/client/service/model/profile/LocalProfileService.java
@@ -1,6 +1,6 @@
 package egl.client.service.model.profile;
 
-import egl.client.model.profile.LocalProfile;
+import egl.client.model.local.profile.LocalProfile;
 import egl.client.repository.profile.LocalProfileRepository;
 import org.springframework.stereotype.Service;
